Docs corrections

See merge request maicos-devel/maicos!268

+++ src/maicos/modules/densityplanar.py
@@ -59,7 +59,7 @@ class DensityPlanar(ProfilePlanarBase):
         unwrap: bool = True,
         bin_method: str = "com",
         output: str = "density.dat",
-        concfreq: float = 0,
+        concfreq: int = 0,
         jitter: float = 0.0,
     ):
         self._locals = locals()
